@@ -1,5 +1,7 @@
-const debug = require('debug')('apicarto');
-const app = require('./app');
+import debug from 'debug';
+import { app } from './app.js';
+//const debug = require('debug')('apicarto');
+//const app = require('./app');
 
 var port = process.env.PORT || 8091;
 app.listen(port, () => {
